// src/app/opmobilitybackend/models/get-town-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface GetTownResponseDto {

  /**
   * Country id
   */
  countryId: number;

  /**
   * Country name
   */
  countryName: string;

  /**
   * Id of the town
   */
  id: number;

  /**
   * Name of the town
   */
  name: string;
}